requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
google-genai>=1.0.0
Pillow>=11.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.16.0
